--help

+++ project2/codebase/rm/rmtest_14.cc
@@ -24,11 +24,12 @@ RC TEST_RM_14(const string &tableName)
 
     RID rid;
     void *returnedData = malloc(200);
+    // memset(returnedData, 0, 200);
 
     // Set up the iterator
     RM_ScanIterator rmsi;
     vector<string> projected_attrs;
-    for (int i = 0; i < attrs.size(); i++){
+    for (int i = 0; i < (int)attrs.size(); i++){
       projected_attrs.push_back(attrs[i].name);
     }
