+++ project2/codebase/rm/test_util.h
@@ -313,7 +313,6 @@ RC createTable(const string &tableName)
     attr.type = TypeInt;
     attr.length = (AttrLength)4;
     attrs.push_back(attr);
-
     RC rc = rm->createTable(tableName, attrs);
     assert(rc == success);
     cout << "****Table Created: " << tableName << " ****" << endl << endl;
@@ -420,7 +419,7 @@ void writeRIDsToDisk(vector<RID> &rids)
 
 	if (ridsFile.is_open()) {
 		ridsFile.seekp(0, ios::beg);
-		for (int i = 0; i < rids.size(); i++) {
+		for (int i = 0; i < (int)rids.size(); i++) {
 			ridsFile.write(reinterpret_cast<const char*>(&rids[i].pageNum),
 					sizeof(unsigned));
 			ridsFile.write(reinterpret_cast<const char*>(&rids[i].slotNum),
@@ -439,7 +438,7 @@ void writeSizesToDisk(vector<int> &sizes)
 
 	if (sizesFile.is_open()) {
 		sizesFile.seekp(0, ios::beg);
-		for (int i = 0; i < sizes.size(); i++) {
+		for (int i = 0; i < (int)sizes.size(); i++) {
 			//cout << sizes[i] << endl;
 			sizesFile.write(reinterpret_cast<const char*>(&sizes[i]),
 					sizeof(int));
